Drawing player sprites
In this version, Jet (player0) and bomber (player1) sprites are drawn with the correct color in the TV pictures. Bomber sprite is stretched accordingly. 

In the TV pictures, check if we are inside the player sprites. If we are, then draw the sprites of bomber and jet. If not, output the blank scanlines.